requirements.txt
Django==3.2
PyJWT>=0.3,<4
psycopg2-binary==2.8
djangorestframework_simplejwt
djoser==2.0.5
Pillow==8.3.2
djangorestframework==3.12.3
python-jose==3.2
django-heroku==0.3.1
gunicorn==19.9.0
whitenoise
markdown     
django-filter  
botocore<1.24.0,>=1.23.14
python-dateutil<3.0.0,>=2.1
urllib3<1.27,>=1.25.4
django-storages==1.12.3
boto3==1.20.14
django-ipware==4.0.0